refactor(container): align components and data with µ-app structure
ref: #MANAGER-18468
