+++ packages/manager/apps/container/src/components/agreements-update-modal/agreementsUpdateModal.constants.ts
