+++ packages/manager/apps/container/src/components/sso-auth-modal/ssoAuthModal.constants.ts
