+++ packages/manager/apps/container/src/components/suggestion-modal/suggestionModal.constants.ts
