+++ packages/manager/apps/container/src/data/api/procedure/procedure.ts
@@ -1,0 +1,10 @@
+import { v6 } from '@ovh-ux/manager-core-api';
+import { Procedure, Procedures } from '@/types/procedure';
+import { PROCEDURE_API_PATH_MAP } from '@/data/api/procedure/procedure.constants';
+
+export const fetchProcedureStatus = async (procedure: Procedures) => {
+  const { data } = await v6.get<Procedure>(
+    `/me/procedure/${PROCEDURE_API_PATH_MAP[procedure]}`,
+  );
+  return data;
+};
