+++ packages/manager/apps/container/src/helpers/procedures/proceduresHelper.ts
@@ -1,5 +1,5 @@
 import { User } from "@ovh-ux/manager-config";
-import { requiredStatusKey } from "@/identity-documents-modal/IdentityDocumentsModal.constants";
+import { requiredStatusKey } from "@/components/identity-documents-modal/identityDocumentsModal.constants";
 
 export const isUserConcernedWithIndiaProcedure = (user: User) => !user.kycValidated;
