--- packages/manager/apps/container/src/hooks/accountUrn/useAccountUrn.tsx
@@ -1,11 +1,0 @@
-import { DefinedInitialDataOptions, useQuery } from '@tanstack/react-query';
-import { fetchAccountUrn } from '@/api/authorizations';
-
-export const useAccountUrn = (
-  options?: Partial<DefinedInitialDataOptions<string>>,
-) =>
-  useQuery({
-    ...options,
-    queryKey: ['account-urn'],
-    queryFn: fetchAccountUrn,
-  });
