--- packages/manager/apps/container/src/hooks/agreements/useAgreements.tsx
@@ -1,13 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-import fetchPendingAgreements from "@/api/agreements";
-
-export const useAgreementsPageNavigationParam = () => ({
-  app: 'billing',
-  path: `#/autorenew/agreements`,
-});
-
-export const usePendingAgreements = (enabled: boolean) => useQuery({
-  queryKey: ['pending-agreements'],
-  queryFn: fetchPendingAgreements,
-  enabled,
-});
